+++ platform/linuxbsd/xkbcommon-so_wrap.c
